move recommendation into protocol_card.py
