+++ canvas_sdk/effects/protocol_card/__init__.py
@@ -1,2 +1,1 @@
-from canvas_sdk.effects.protocol_card.protocol_card import ProtocolCard
-from canvas_sdk.effects.protocol_card.recommendation import Recommendation
+from canvas_sdk.effects.protocol_card.protocol_card import ProtocolCard, Recommendation
